refactor(agent): split AgentPanel.tsx into focused files
Move the self-contained leaf components out of the 2.5k-line AgentPanel
into siblings: AgentSidebar, AgentChatInput (input + slash popup +
compact/stats popovers), AgentSettingsView (settings + marketplace), and
ModelPicker. AgentPanel drops from 2524 to 1003 lines and is now just the
orchestrator. No behavior change.
